+++ src/commands/wizard.ts
@@ -39,8 +39,9 @@ export const assembliesDescriptions: Record<Assembly, string> = {
     [Assembly.dual]: 'Dual Node',
     [Assembly.peer]: 'Peer Node',
     [Assembly.api]: 'Api Node',
-    [Assembly.demo]: 'Demo Node',
+    [Assembly.demo]: 'Demo Node. A dual node that includes a Faucet and Explorer.',
     [Assembly.multinode]: 'Multinode Node. A docker compose that includes one api, one rest and two peers.',
+    [Assembly.services]: 'Services. A docker compose with just Faucet and Explorer.',
 };
 
 export enum HttpsOption {